tests/wpt/mozilla/tests/mozilla/document_domain.html

<!doctype html>
<html>
    <head>
        <title></title>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script>
            test(function() {
                assert_equals(typeof document.domain, "string", "document.domain is a string");
                assert_not_equals(document.domain, "", "document.domain is not empty");
            }, "sanity checks");

            test(function() {
                var doc = new Document();
                assert_equals(doc.domain, document.domain, "new document has same domain");
            }, "new document");

            test(function() {
                assert_equals(document.domain, window.location.hostname, "equals location.hostname");
            }, "current document");
        </script>
    </head>
    <body>
    </body>
</html>
